+++ CHANGELOG.md
@@ -34,6 +34,7 @@
 - Resolve macro usages in the header parameter for attribute sub-type [#1362](https://github.com/epam/sap-commerce-intellij-idea-plugin/pull/1362)
 - Added code completion and reference resolution for `&DocId` usages [#1364](https://github.com/epam/sap-commerce-intellij-idea-plugin/pull/1364)
 - Added rename capabilities for Document ID `&DocId` elements [#1365](https://github.com/epam/sap-commerce-intellij-idea-plugin/pull/1365)
+- Show variants for Document ID `&DocId` usage [#1366](https://github.com/epam/sap-commerce-intellij-idea-plugin/pull/1366)
 
 ### `Terminal` integration
 - Added support of the `Terminal` plugin and set default directory to the project root for new Terminal Windows [#1342](https://github.com/epam/sap-commerce-intellij-idea-plugin/pull/1342)